Make xpSinceReward config an int again
Same reason as previous commit

+++ src/main/java/com/cworldender/PersonalCurrencyTrackerConfig.java
@@ -244,7 +244,7 @@ public interface PersonalCurrencyTrackerConfig extends Config
 		description = "",
 		hidden = true
 	)
-	default long xpSinceReward(){ return 0; }
+	default int xpSinceReward(){ return 0; }
 
 	@ConfigItem(
 		keyName = "xpSinceReward",
